[sailjaild] React to application changes. JB#54096 OMP#JOLLA-108
Adjust invocation checks so that they can be used to check if invocation
has become resolvable and send a reply if that's the case. Handle error
cases more accurately.

Add change notifier for changed applications or application settings
changes to prompter. If some change resolves current invocation or items
in the queue they are replied immediately. Also cancel windowprompt when
current invocation is resolved.

Signed-off-by: Tomi Leppänen <tomi.leppanen@jolla.com>